requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.23.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pypdf>=3.17.0
